refactor(SSP-26): recorrecting between two sites
